requirements.txt
streamlit
opencv-python-headless
mediapipe
numpy
scipy
Pillow
